requirements.txt
kafka-python
Faker

